requirements.txt
dnspython==2.2.1
requests~=2.28.1